Revert "logics 폴더와 prompts 폴더 내용물 및 경로 정리"
This reverts commit f522bd46d6de3dbcfca46b4ecf749f2d6829793f.
